+++ src/ui/routes/Users/UserTable.tsx
@@ -15,6 +15,7 @@ const useTranslation = makeTranslationHook({
       username: "Username",
       displayname: "Displayname",
       email: "Email",
+      class: "Class",
       role: "Role",
     },
   },
@@ -23,6 +24,7 @@ const useTranslation = makeTranslationHook({
       username: "Nutzername",
       displayname: "Displayname",
       email: "Email",
+      class: "Klasse / Stufe",
       role: "Rolle",
     },
   },
@@ -64,6 +66,13 @@ export const UserTable: React.FunctionComponent<UserTableOwnProps> = (
           extract: (user) => user.get("email"),
           options: {
             filter: false,
+          },
+        },
+        {
+          name: lang.headers.class,
+          extract: (user) => user.get("class"),
+          options: {
+            filter: true,
           },
         },
         {